feat(handler): refactor and add TodoHandler templates for Gin, Chi, and Echo frameworks
